Load exercises from assessment queue in assessment mode
Summary:
Instead of getting exercises from the bottomless queue, this will tell the assessment queue the answer to the current problem done and have it serve up the next exercise when it is ready.

This changeset also moves the enabling of the check answer button to the end of make problem rather than the beginning of it.

Test Plan:
Go to the homepage with a new user in athena and start the pretest.  Make sure that three different exercises are loaded for the current problem and its followups.  When answering check that it is not loading any more exercises until the response from the server says the next upcoming exercises.

Load up exercises in learning tasks and from the library to make sure questions can be answered there without the answer button doing anything differently from before.

Reviewers: chris, alpert

Reviewed By: alpert

CC: jace, benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2671