Don't accept maxError if we don't specify inexact
Summary:
Before, if inexact was left as "required" but a maxError was still
provided, inexact answers could be provided and would be accepted. This fixes
the problem by changing maxError if the answer isn't supposed to be exact. I
could also try to change the logic down below to account for this, but I think
that this is the easiest way to fix.

Test Plan:
Add `data-max-error="100"` to the solution of an exercise, like
addition_1. Make sure that an answer 100 away from the correct answer isn't
still accepted.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2813