Fix i18n issues in triangle_inequality_theorem
Summary:
This exercise was doing sentence building, as well as doing math on
`localeToFixed`ed numbers, which was causing errors. This fixes all of that.
Closes T1133.

Test Plan:
`make i18n`. Open the problem with lang=en. Ensure that all numbers
are shown with only one decimal place everywhere. Make sure that the math in
the hints makes sense (i.e. no rounding errors). Open the problem with lang=es.
Ensure that all of the strings are translated, that all the numbers use a comma
instead of a decimal point. Ensure that there are no `NaN`s, and that the
correct answer is accepted.

Reviewers: john

Reviewed By: john

CC: aatashparikh, alpert

Maniphest Tasks: T1133

Differential Revision: http://phabricator.khanacademy.org/D2882