Fix MathJax updating to be faster
Summary: Turns out that the issue with the slowness was that we weren't calling the MathJax queue callback in the same event loop. Now all the exercises mentioned in T1178 work properly, with the exception of one_step_equation_intuition which has nothing to do with graphie labels.

Test Plan: Interact with each of the exercises in T1178 and see that the MathJax doesn't flicker when rerendering.

Reviewers: eater, emily

Reviewed By: eater

Maniphest Tasks: T1178

Differential Revision: http://phabricator.khanacademy.org/D2919