Make less awkward i18n in lcm_gcd_word problems
Summary:
By using a bunch of nested ifs, recombine some of the sentences
originally split up for i18n, to make the sentences work better. Also, use a
table for styling in one section so that it appears correctly in the exercise
regardless of width.

Test Plan:
Open
`/exercises/least_common_multiple_and_greatest_common_divisor_word_problems.html`
in local mode. See that the hints in all the problem types have sentences like
"1 set with 40 pencils and 20 folders" and not "There is 1 set. It has 40
pencils. It also has 20 folders". See that the comparison between the two
different groups are next to each other.

Reviewers: alpert, eater

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D3002