Fix radio for none of the above
Summary:
When "none of the above" is selected, and we're missing a single
choice, add in "none of the above" as an answer, instead of dying. Fixes T1288.

Test Plan:
Open a question with 4 choices and data-show=5, data-none="true". See that it
adds back in "none of the above" as an answer. Also, see that the tests work.

Reviewers: alpert

Reviewed By: alpert

Maniphest Tasks: T1288

Differential Revision: http://phabricator.khanacademy.org/D3374