Fix bugs with example math not showing up
Summary:
When we transfer the examples from the answer-type to the qtip,
they're stored as strings. So, when there's math in them, we need to be sure
that it's unrendered math. Then, when we add it back into the qtip, we can't
re-render right before calling qtip, because we have to remove it then
immediately, and so MathJax will fail to render the math. Instead, we only
re-render it when the qtip itself is rendered, so Mathjax doesn't get confused.

Test Plan:
Make sure /exercise/completing_the_square_2 works in both athena and
non-athena with KaTeX disabled and with KaTeX enabled.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D3670