Mark an answer empty only if all fields are empty
Summary:
Previously, we were sometimes inconsistent about this but we made a small effort to mark things as empty if anything was empty. I'm changing this so that it's marked empty only if everything is empty.

Note my changes to the graphing exercises are not the result of a regression; they returned false before and now return empty on an unmoved graph.

Test Plan: Ran unit tests (unit tests are great!), tested every exercise that has a checkbox.

Reviewers: emily, eater

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D3674