Warn about using X for * in expressions
Warns by using expression parser error message box.
Doesn't count the answer as wrong, ignores it instead.

Has a test for using x as mul defined in two places
as there doesn't seem to be a good place to put a
function that's going to be shared  between the
validator and setup code.