Get rid of cardinal() everywhere escept the exercise testing cardinality.
Summary:
In most places, I replaced cardinal with the simpler
cardinalThrough20() function, which shows a cardinal number if n is
between 0 and 20, and the digits of n otherwise.  In almost all uses,
n is provably between 0 and 20 so this should give mostly the same
results as before.

In a few cases, such as some of the age_word_problems, n can sometimes
be > 20.  However, in those exercises, we seem to be mixing cardinal
numbers with decimal numbers somewhat indisciminately anyway, so I
doubt anyone will notice the change.

Task T1060

Get rid of ordinal() as well, replacing it with ordinalThrough20().

This is enough to cover all the uses of ordinal to date.

Test Plan:
Ran python -mSimpleHTTPServer and visited
   http://localhost:8000/exercises/age_word_problems.html
and likewise for all the exercises in this commit, and looked at the
text and hints to make sure they read ok.  I reloaded until the
exercise(s) with the cardinal change came up.

Ran python -mSimpleHTTPServer and visited
   localhost:8000/exercises/integer_sums.html
and likewise for all the other exercises, and read the text and hints
to make sure the ordinal numbers were displayed correctly.

Reviewers: alpert, eater

Reviewed By: alpert

Maniphest Tasks: T1060

Differential Revision: http://phabricator.khanacademy.org/D3926