Metacognitive experiment.
Summary:
These are the modified exercise files for Joseph's proposed metacognitive experiment,
where students are asked to engage in thought about what they need to do to answer the
question and/or why they need to do it.  There will be a corresponding checkin for the
webapp change to select these exercise files by GAE/Bingo condition.  The trailing
portion of the filenames describe which experimental condition each file corresponds to.

no extension -- Current exercise.
'Control' -- Only the what question.
'SelfRegulatingThinking' -- Introspective versions of the header questions.
'ReflectingOnMeaning' -- Inspective versions of the header questions.

'NoTextBox' -- There are no fields for the student to enter information in the problem hints.
'TextBox' -- There are fields for the student to enter information in the problem hints. This
contains custom validation code, which calls the standard validator, so that the contents of
the text fields are saved along with the answer.

This is a followup to the code review at http://phabricator.khanacademy.org/D2096 .  All of
the complaints in that code review have hopefully been addressed, and the code has otherwise
changed significantly.  I have only limited knowledge of javascript/html though, so I
apologize in advance for what I'm sure will be some ugly parts.

The code was written by Joseph, by me, and by an undergrad working with Joseph (whose name
I don't remember and don't have in email).

Test Plan:
I've made sure these exercise load correctly, that the answer validation works correctly for
the custom validator case, and that the timeline works correctly for them (even in the case
where the student is in one experimental condition and the coach is in another experimental
condition, and in the case that the timeline was saved prior to the start of the experiment).

Reviewers: jace, eater

Reviewed By: jace

CC: alpert, chris

Differential Revision: http://phabricator.khanacademy.org/D2809