Lint fixes for the chemistry exercises.
Summary:
Whew!  Having to translate 'mole' everywhere was a big task.

I improved the linter a bit to make it better recognize non-alpha text
inside \text{...}.  I also updated some comments and TODOs.

Test Plan:
Ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/ideal_stoichiometry.html
and
   http://localhost:8000/exercises/limiting_reagent_stoichiometry.html

and read the exercise and all the hints and all the sub-hints, and
verified they read reasonably in English.

Reviewers: eater, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D4103