Simplify pack.rb a bit more by not supporting streaming-stdin anymore.
Summary:
Nobody uses the functionality where we cat lots of input files in
stdin, and then expect pack.rb to separate them on <!docstring> tags.
So I got rid of that, and now assume that if you pass something in on
stdin, it's a single file to be processed.

This lets me unify the stdin (filter) case and the argv case, yielding
much more readable code.

Test Plan:
Ran the following:
   make exercises
   make exercises  # a no-op this time
   ruby khan-exercises/build/pack.rb `pwd`/khan-exercises/exercises/absolute_value.html
   ruby khan-exercises/build/pack.rb `pwd`/khan-exercises/exercises/absolute_value.html::/var/tmp/a

The make commands succeeded, the second time rather quickly (due to
filemod_db), and the 3rd command correctly emitted packed content to
stdout, while the 4th correctly emitted it to the file.

Reviewers: eater

Reviewed By: eater

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4592