Listen + trigger events to show prereqs in focus mode
Summary: This diff corresponds to another in webapp, and I'm not entirely sure which order to land what and how. But, I'm sure it'll be clear soon! In this diff, we listen to clicks on the "Help me learn this" button and trigger an event that Dashboard.FocusMode will listen for.

Test Plan:
Log in as a user who passes the show_prereqs_in_focus_mode gandalf bridge
Open a practice task in focus mode
See the help me learn this button:
https://s3.amazonaws.com/uploads.hipchat.com/6574/17223/fl7tzicnwomsxnj/upload.png

Click on it and see that the button is disabled and the prereqs appear to the left:
https://s3.amazonaws.com/uploads.hipchat.com/6574/17223/3u6hnjav6tind5t/upload.png

Click on a prereq and see that the prereq's title is highlight as selected and the "Help me learn this button" is enabled.
Click the bottommost in the list, which corresponds to the originally opened task
See that the "Help me" button is disabled (since its prereqs are already shown)

Click back to the prereq
Click on this prereq-task's help button and see that the list to the left reflects the prereq-task's prereqs.

Also go through problems and verify that the help me button doesn't change

Open addition_1 in focus mode
Verify no help me button appears (addition_1 has no prereqs)

Open an exercise in library mode
Verify no help me button appears

Reviewers: alpert

Reviewed By: alpert

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4621