Make bounding of MovablePoint optional; fix #92677
The congruency_postulates exercise draws the triangle in such a way that it might be off-screen, then once it's drawn it translates it so it's visible. By forcing the movablePoints to be onscreen at all times, the triangles were sometimes being drawn slightly misshapen leading to an unanswerable question. Since it's okay if some of the points are off-screen as the user drags it around (as long as some of the triangle remains on screen, which is already handled properly) I added an option to disable bounding the point.

I also removed references to the old "fixed" property in MovablePoint since it has been replaced by "constraints.fixed".

Test plan: Ensure that http://localhost:8000/exercises/congruency_postulates.html?seed=36&problem=0&debug is answerable. Also `awk fixed:` to verify that the `fixed` property of movablePoint isn't used anywhere (everything use constraints.fixed)

Auditors: jack