Add a small tolerance to floating point inexact answers
Summary:
Previously, the check for whether an inexact numerical answer
was correct was

    abs(guess - correct) < maxError

This lead to unfortunate circumstances such as:
2.3 isn't correct for 2.45 +/- 0.15
25 isn't correct for 24.5 +/- 0.5

This change always adds a small near-epsilon number to maxError,
which was already being added for exact answers.

Test Plan:
Make a question accepting inexact 2.45 with maxError 0.15. Verify
that 2.3 is rejected prior to this change, but accepted afterwards.

Reviewers: alex, eater

Reviewed By: eater

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4735