I18nize the decimal point in the calculator
Summary:
Some languages have , instead of . as the decimal separator.  Together with a
one liner in webapp to add the classname, this review will change the text of
the button and make sure the input and output is converted.

Test Plan:
Load an exercise on es.khanacademy.dev
Do the following in the console:
$("#calculator").show();
Try 1,5*3
Make sure it equals 4,5

Reviewers: alpert

Reviewed By: alpert

CC: jessica

Differential Revision: http://phabricator.khanacademy.org/D4923