Update template with classes in addition to ids.
Due to the constraints of perseus one (perseus 360?) (one-page-app stuff
- ask me if you're interested) I sometimes have to hide the page body
and show a perseus renderer. That renderer breaks because it gets
workarea, solutionarea, and hintsarea by id... but those ids already
exist on the page and they're not the elements we want.

But what if there was a way to name elements on the page in such a way
that there could be multiple elements with the same label? I scoured the
html spec and discovered a little-know but well supported feature called
[classes](http://www.w3.org/TR/html5/dom.html#classes).

It's risky, but my plan is to select dom nodes using these "classes"
instead of ids.

Test Plan:
Ack for workarea, solutionarea, and hintsarea - determine these class
names don't already exist.

Auditors: alpert
Cc: alex