workspace (tex-pad?) idea
The general idea is to allow people who may not have a touchscreen or
paper/pencil to do math on the screen via tex, or a pseudo-form of it.

Basically copies scratchpad/expression code, with a few changes to
handle multiple lines, and clear correctly. Also allows toggle between
tex and non-tex. By "non-tex", I mean that it is parsed by KAS, and
allows (only) parentheses instead of braces for expressions, doesn't
require a backslash in front of operators like `sin` , among other
differences.

Does not heed the "change in both locations" in khan-exercise.html
because I couldn't find
"javascript/exercises-package/problem-template.handlebars"