Add clues for common mistakes/misconceptions
Provides a clue if the user inputs an answer that is a common
mistake/misconception (arithmetic error, misunderstanding, wrong
format). Gives the exercise-creator the ability to add data-clues
property to solution in form `'{"guess":"message"}'` with an option to
pass an array as the value: `'{"guess":["message",
"count_as_empty"]}'`, where count_as_empty is set to true to prevent
the guess from being marked incorrect (a more lenient approach).