Put fallback on set-sol tag instead of on template
I broke this a while ago... not convinced this is how it should be but at least this way it works and you don't need to type `0`.

Fixes #114487.

Test Plan:
Answered both of multiplying_and_dividing_rational_expressions_4.html?seed=48&problem=B-over-A-and-C-over-BC and multiplying_and_dividing_rational_expressions_4.html?seed=49&problem=B-over-A-and-C-over-BC successfully.

Auditors: eater