Don't show unrelated videos
For a while, I've been seeing a steady stream of reports of people seeing the wrong related videos. At least one culprit seems to be that we weren't properly clearing out related videos in mastery challenges for problems where the user was already mastery2 or mastery3. In that case we don't want to show them any video, but instead we were leaving behind the videos from the previous problem.

Test Plan: It was painful, but crank down the MIN_TIME_ELAPSED_FOR_MASTERY_INCREASE in missions/recommender.py and do lots of practice and mastery tasks until randomly getting a mastery challenge with an exercise in practiced or mastery1 followed by an exercise in mastery2 or mastery3 within the same MC. Some logging was helpful to confirm this condition. Before this change, the second exercise would retain the related videos from the first. After this change (and repeating the repro steps again), the videos disappeared when the second exercise was rendered.

Auditors: benkomalo