Make click-to-add-point work in IE8
In IE, the VML "canvas" itself wasn't getting mouse events. It seems you only get mouse events on objects that are filled (or if you carefully click the ouline of an unfilled object). Otherwise, the event gets passed to whatever's under the canvas (in our case, sometimes an image). If there's no image, then the canvas gets the event after all. Gotta love IE. The result is that click-to-add points weren't working when a background image was set. This commit adds a big filled in (yet invisible) rectangle and hooks the click handler up to it. This is actually consistent with the other invisible shapes that roam around graphie.mouseLayer, so it doesn't feel like too much of a hack.

Test Plan: Use testrender.html to create a scatterplot in IE8 w/ background image. Click to add points.

Auditors: jack