Add some rounding to deal with machine imprecision
This also removes the fractionSimplify function which I'm pretty sure does the same thing as fractionReduce.

Test Plan: Lots of refreshing and viewing all hints in multiplying_rational_numbers to look for things that weren't rounded properly. Also in the console, verify that KhanUtil.fractionReduce(4, 2) returns just "2". The "defraction" flag to fraction() is supposed to do this.

Auditors: peter