Rename or delete some not-deployed exercises files
None of these files are live on the site. Some will soon be deployed with a name different than the filename so I'm taking the opportunity to rename the file before it's deployed and it's still convenient to do so.

Two of these files we don't want to deploy, so I'm deleting them from the repo.

Test Plan: none

Auditors: elizabeth, joel