Save deg/rad mode in localstorage.
Summary: Default to DEG instead of RAD, but save the mode in localstorage if the user switches it.

Test Plan: Load the law_of_cosines exercise and note that it starts in DEG mode, change modes and note that a localstorage entry is created. Refresh the page and note that it remembers what state it was in.

Reviewers: joel

Reviewed By: joel

CC: alpert

Maniphest Tasks: T1987

Differential Revision: http://phabricator.khanacademy.org/D5542