Divert perseus issue reports to JIRA instead of GitHub
Testing out a different platform for issue tracking that will hopefully make it easier to get issue reports directly to content creators. We'll see how this goes.

Test Plan: With the corresponding changes in webapp, report an issue on a khan-exercises problem and a perseus problem. In each case, see that the issue gets created in the appropriate tracker.

Auditors: alex