Add a message on "Check Answer" when empty
Summary:
Displays "It looks like you haven't answered all of the question yet."
to the student if they click "Check Answer" without having filled out
all of the question.

Hopefully this reduces the number of confused students who think the
website is broken (and the number of bug reports sent for not filling
in the whole question).

Test Plan:
 * python -m SimpleHTTPServer 7000
 * visit http://localhost:7000/exercises/factoring_linear_binomials.html
 * press "Check Answer" without filling anything out
 * answer the question and make sure it still works

Reviewers: eater

Reviewed By: eater

CC: alpert, alex

Differential Revision: http://phabricator.khanacademy.org/D5898