MovableAngle: Add snapOffsetDeg parameter
Summary:
Adds a snap offset parameter, so that your angle rays can snap to,
for example:
    6, 11, 16, 21 with a snapDegrees of 5 and a snapOffsetDeg of 1
instead of only
    5, 10, 15, 20 with a snapDegrees of 5

Note: The angle between the rays must still be a multiple of
snapDegrees. I tried changing that to allow for more flexibility,
but it just ended up making the snap interaction clumsy and confusing
when paired with the angle wraparound behaviour.

Test Plan: Used in Perseus' interactive-graph angle widget

Reviewers: alex, justinjaffray

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6127