Ask user to refresh when page load fails
Summary: There were a couple options (retry ajax request, ask user to refresh, retry than ask), but I'd like to see (a) how prevalent the issue is on a single day

Test Plan: It's because of the testing that I have the if statement within the done. The get/ajax requests wouldn't fail, just succeed with a 404 page. Once I've confirmed this works in production, we can take out the autoSubmitIssue and if-statement in the done

Reviewers: alpert

Reviewed By: alpert

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6450