Add Travis CI
Summary:
This commit adds a file for https://travis-ci.org/ which will alert the
"Content Tools" room if there is a commit which breaks the tests.

If you have a preference for a different room let me know though.

After this gets committed we still need to flip the switch on the Travis
website before anything will happen

Test Plan:
  * It worked when I did it on a fork
    * https://travis-ci.org/justinj/khan-exercises

Reviewers: eater

Reviewed By: eater

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6570