Log "unclean" HTML
Summary: cleanHTML is basically called on the .text() property of vars, which shouldn't contain HTML entities. If we don't get any issues I'll remove it.

Test Plan: Added `"&amp;gt;"` to one of the vars in addition_1 and changed autoSubmitIssue to log the description after adding the stack; then loaded adddition_1 and saw the description followed by the stack trace.

Reviewers: jack, eater

Reviewed By: eater

Differential Revision: http://phabricator.khanacademy.org/D7145