Fix wrong answers in sets being marked as right
Apparently if an answer type within a set is wrong, but returns a message, the set logic will mark it as correct (but still show the message saying it's almost correct).

Test Plan: Run the unit tests. Also try entering a lower-case answer to the `vertical` problem type of `exploring_angle_pairs`

Auditors: peter