Remove `debugger;`
Summary:
I'm not really sure why this doesn't seem to be a problem in the live
site - it goes into debug mode serving the exercises locally, but
doesn't when I open up the exercise on webapp.

Either way, I don't see a good reason to keep it (and it keeps showing
up in my greps).

Test Plan:
* Open http://localhost:8000/exercises/triangle_inequality_theorem.html
* Refresh a few times, should not get debugger.

Reviewers: eater

Reviewed By: eater

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D7559