Fix toFraction
Summary:
Evidentally 283.33 % 1 = .3299999999999841

Also, using our previous method, 1/1000 took 1000 steps to get to.

This adds (a) a fix to that, (b) a worst case running time of 100 steps

Anything beyond that is a useless fraction.

It accounts for all fractions where n < d < 100
```
for (var x = 1; x < 100; x++) {
  for (var y = x+1; y < 100; y++) {
    var z = KhanUtil.toFraction(x/y);
    if(x/y !== z[0]/z[1]) {
      console.log(x, y, z);
    }
  }
}
```

The first non-exact fraction is for d >= 141

Test Plan:
Try out:
(1) 283.333333333333333
(2) 283.33
(3) .0001

Reviewers: brianmerlob

Reviewed By: brianmerlob

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D7519