log subhint usage
Summary: Added code to log when subhints are clicked on and expanded.  This will be used in the header text intervention experiment to see when users interact with dropdown text.  This will not work without the corresponding webapp checkin, at http://phabricator.khanacademy.org/D6914 .

Test Plan: Ran it on local appserver.

Reviewers: jace, alpert, eliana

Reviewed By: alpert, eliana

Subscribers: alpert, mattfisher, mattfaus

Differential Revision: http://phabricator.khanacademy.org/D6913