Use Exercises.getTaskId() function in interface.js to find learning task ID with no learning task
Summary: The iOS web view doesn't have a learningTask on the Exercises object, so couldn't post that parameter back to the server on making an exercise attempt. Since it only needs the taskId, I'm creating a function on the iOS Exercises object that just returns that piece. For now, to make minimal changes to the web app, I'm leaving the current structure in place and just adding this additional path to the data we need to complete the request.

Test Plan: Do some exercises on the iOS app and in the web app and make sure they advance to the next one and handle attempts appropriately.

Reviewers: mgp, ben

Reviewed By: mgp, ben

Differential Revision: http://phabricator.khanacademy.org/D8297