Make sure LocalStore exists before trying to access it for the scratchpads
Summary:
This section assumes a LocalStore is set up, which isn't currently true on the iOS app (and may not be in the future, either).

Checking its existence before accessing it avoids an exception.

Test Plan: Ran khan-exercises in both the iOS and web apps; scratchpads work in the web still, and the iOS app doesn't throw an exception there.

Reviewers: mgp, ben

Reviewed By: mgp, ben

Subscribers: marcos

Differential Revision: http://phabricator.khanacademy.org/D8295