Make sure oauth ajax requests from exercises are going to ka.org
Summary: We'd tried appending https://ka.org in $.ajax beforeStart, but since that happens after oauth is set up, authentication failed. This adds ka.org to the requested URL if the URL is missing a domain and is trying to access the API (without this change, it defaulted to the window's location protocol, which is file:// in the iOS app).

Test Plan: Logged in & did exercises in the iOS & web apps; progress posted to the server successfully!

Reviewers: mgp, ben

Reviewed By: ben

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D8321