fixes lint error for global oauth var in interface.js
summary: oauth is defined only for specific projects that include it in the global namespace to perform authenticated calls to the api. In this case, interface.js assumed that webview-bridge was available and was setting the `oauth` object in the global namespace, but this isn't always the case so the linter complained.

test plan: rebuilt ipad app and successfully saw a 200 OK when posting to /attempt

auditors: laura, ben