Limit the number of times we try and remake a bad problem.
Summary:
While investigating https://github.com/Khan/khan-exercises/issues/161349
it turned out that MakeProblem kept failing, and we kept trying, until
we hit the maximum call-stack size.  In the meantime (many minutes on my
computer) the browser froze.  This change is intended to break out of this
call stack loop when this happens again.

Test Plan:
http://localhost:8080/student-res/algebra1-brushup/quad-eqns-brushups/e/solutions_to_quadratic_equations?lang=es
See that after a second a warning gets spit out into the console.

Reviewers: eater

Reviewed By: eater

Subscribers: jack, alpert

Differential Revision: http://phabricator.khanacademy.org/D8547