Better handle failed validator setup
Summary:
Before when hitting http://exercises.ka.local/exercises/trig_addition_identities.html?debug&problem=cosAdd&seed=125 it would retry since seed cosAdd/125 doesn't generate enough wrong answers. But it would retry crazy seeds like 12917421, 70115338, and 201142212 before giving up because 3 attempts.

Increasing the retry count alone is enough to cause a valid problem to be generated (seed 207996478 in this case). But it didn't make any sense to me why we were passing randomSeed (which is the prng's seed, not the problem seed) to makeProblem. Instead it seems like we ought to just call makeProblem with the same parameters it was originally called with.

As far as I know, seedOverride is only passed when rendering the timeline, so in theory we'd never get stuck in a loop rendering the same seed over and over unless the seed was valid when the student first did it, but is now invalid when looking at it in the timeline. In that case, the 100 tries will all fail and we won't render anything. Not sure what else we could do in that case though--by definition we can't render the problem the user saw.

Test Plan:
1. Hit http://exercises.ka.local/exercises/trig_addition_identities.html?debug&problem=cosAdd&seed=125
2. Notice that it gives me seed #21 since 125 doesn't generate enough wrong answers.

Reviewers: jack, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D9146